Extract authentication tests into a separate module
